classify char
