+++ app/src/main/java/com/example/aganada/views/WordView.kt
@@ -11,6 +11,7 @@ import android.view.View
 import android.graphics.PointF
 import com.example.aganada.R
 import com.google.mlkit.vision.digitalink.Ink
+import kr.bydelta.koala.dissembleHangul
 import kotlin.math.atan2
 import kotlin.math.pow
 import kotlin.math.sqrt
@@ -52,6 +53,34 @@ class WordView @JvmOverloads constructor(
     private var eraserPath: Path? = null
     private var eraserPoint: PointF? = null
 
+    private val onset = listOf(Char(0x1100), Char(0x1101), Char(0x1102),
+        Char(0x1103), Char(0x1104), Char(0x1105), Char(0x1106),
+        Char(0x1107), Char(0x1108), Char(0x1109), Char(0x110A),
+        Char(0x110B), Char(0x110C), Char(0x110D), Char(0x110E),
+        Char(0x110F), Char(0x1110), Char(0x1111), Char(0x1112))
+    //0x1161..0x1175
+    private val nucleus = arrayListOf(Char(0x1161), Char(0x1162), Char(0x1163),
+        Char(0x1164), Char(0x1165), Char(0x1166), Char(0x1167),
+        Char(0x1168), Char(0x1169), Char(0x116A), Char(0x116B),
+        Char(0x116C), Char(0x116D), Char(0x116E), Char(0x116F),
+        Char(0x1170), Char(0x1171), Char(0x1172), Char(0x1173),
+        Char(0x1174), Char(0x1175))
+
+    private val horizonNucleus = arrayListOf(Char(0x1169), Char(0x116D),
+        Char(0x116E), Char(0x1172), Char(0x1173))
+    //0x11A8..0x11C2
+    private val coda = arrayListOf(null, Char(0x11A8), Char(0x11A9), Char(0x11AA),
+        Char(0x11AB), Char(0x11AC), Char(0x11AD), Char(0x11AE),
+        Char(0x11AF), Char(0x11B0), Char(0x11B1), Char(0x11B2),
+        Char(0x11B3), Char(0x11B4), Char(0x11B5), Char(0x11B6),
+        Char(0x11B7), Char(0x11B8), Char(0x11B9), Char(0x11BA),
+        Char(0x11BB), Char(0x11BC), Char(0x11BD), Char(0x11BE),
+        Char(0x11BF), Char(0x11C0), Char(0x11C1), Char(0x11C2))
+
+    // 초성, 중성, 종성 획수
+    private val onsetStrokes = listOf(1, 2, 1, 2, 4, 3, 3, 4, 8, 2, 4, 1, 2, 4, 3, 2, 3, 4, 3)
+    private val nucleusStrokes = arrayListOf(2, 3, 3, 4, 2, 3, 3, 4, 2, 4, 5, 3, 3, 2, 4, 5, 3, 3, 1, 2, 1)
+    private val codaStrokes = arrayListOf(0, 1, 2, 3, 1, 3, 4, 2, 3, 4, 6, 7, 5, 6, 7, 6, 3, 4, 6, 2, 4, 1, 2, 3, 2, 3, 4, 3)
 
     init {
         for (c in 'A'..'Z') {
@@ -221,10 +250,17 @@ class WordView @JvmOverloads constructor(
             canvas.drawText(word, textX, textY, paint)
 
             Log.d("sunyoung", "$textX $textWidth")
+            val stroke = strokes(word)
+            for(i in word.indices){
+                println()
+                println(stroke[i])
+            }
             paint.textSize = 70f
-            canvas.drawText("1",
-                (textX - textWidth*2).toFloat(), (centerY - textSpace * 0.75).toFloat(), paint)
-            canvas.drawText("2", textX, (centerY + textSpace * 0.75).toFloat(), paint)
+            canvas.drawText(
+                "1",
+                (textX - textWidth / 2), (centerY - textSpace * 0.5).toFloat(), paint
+            )
+            canvas.drawText("2", textX, (centerY + textSpace * 0.5).toFloat(), paint)
         }
 
         /* draw line */
@@ -250,6 +286,31 @@ class WordView @JvmOverloads constructor(
                 }
             }
         }
+    }
+
+    fun strokes(label: String): IntArray {
+        var strList = IntArray(label.length) { 0 }
+        for (i in label.indices) {
+            if (label[i].dissembleHangul() == null) {
+                strList[i] = 0
+                continue
+            }
+            val char = label[i].dissembleHangul()?.toList() as List<*>
+            if(char[2]==null){
+                if(char[1] in horizonNucleus) {
+                    strList[i] = 1
+                } else {
+                    strList[i] = 2
+                }
+            } else {
+                if(char[1] in horizonNucleus) {
+                    strList[i] = 3
+                } else {
+                    strList[i] = 4
+                }
+            }
+        }
+        return strList
     }
 
     inner class PathData(x: Float, y: Float) {
